+++ src/screens/product-page.jsx
@@ -33,6 +33,14 @@ export function ProductsPage() {
 
   const deleteProduct = async (productId) => {
     try {
+      const productToDelete = products.find(
+        (product) => product.id === productId
+      );
+      if (productToDelete && productToDelete.category) {
+        console.log("Erro ao excluir, categoria vinculada.");
+        return;
+      }
+
       await productService.deleteProduct(productId);
       setProducts(products.filter((product) => product.id !== productId));
     } catch (error) {
@@ -40,12 +48,28 @@ export function ProductsPage() {
     }
   };
 
+  const updateProduct = async (productId, updatedData) => {
+    try {
+      await productService.updateProduct(productId, updatedData);
+      const updatedProducts = products.map((product) =>
+        product.id === productId ? { ...product, ...updatedData } : product
+      );
+      setProducts(updatedProducts);
+    } catch (error) {
+      console.error("Erro ao atualizar produto:", error);
+    }
+  };
+
   return (
-    <main className="h-screen flex-col px-12 bg-white text-zinc-950 flex items-center justify-center">
+    <main className="h-screen flex-col px-12 bg-white text-zinc-950 flex items-center justify-start">
       <Card>
         <ProductForm onSubmit={createProduct} />
       </Card>
-      <ProductTable products={products} deleteProduct={deleteProduct} />
+      <ProductTable
+        products={products}
+        deleteProduct={deleteProduct}
+        updateProduct={updateProduct}
+      />
     </main>
   );
 }
